feat: Add BedTime Calculator style: Change App theme
Bedtime Calculator is added. You can calculate when you have to go to sleep to wake up at when you want to wake up or vice versa.
Also theme of app is changed. There are blue and navy colors. 'Start' and 'Rest' buttons of Timer Activity is merged to Timer. Convert between workout and rest through just tap the timer.